+++ Source/System/Constants.h
@@ -220,6 +220,9 @@ namespace RTE {
 		INPUT_WEAPON_DROP,
 		INPUT_WEAPON_RELOAD,
 		INPUT_WEAPON_PRIMARY_HOTKEY,
+		INPUT_WEAPON_AUXILIARY_HOTKEY,
+		INPUT_ACTOR_PRIMARY_HOTKEY,
+		INPUT_ACTOR_AUXILIARY_HOTKEY,
 		INPUT_START,
 		INPUT_BACK,
 		INPUT_R_UP,
@@ -255,6 +258,9 @@ namespace RTE {
 	    "Drop Device", // INPUT_WEAPON_DROP
 	    "Reload Weapon", // INPUT_WEAPON_RELOAD
 	    "Primary Weapon Hotkey", // INPUT_WEAPON_PRIMARY_HOTKEY
+	    "Auxiliary Weapon Hotkey", // INPUT_WEAPON_AUXILIARY_HOTKEY
+		"Primary Actor Hotkey", // INPUT_ACTOR_PRIMARY_HOTKEY
+		"Auxiliary Actor Hotkey", // INPUT_ACTOR_AUXILIARY_HOTKEY
 	    "Start", // INPUT_START
 	    "Back", // INPUT_BACK
 	    "Analog Aim Up", // INPUT_R_UP